trying to get o2 rebuilt after moving trapsimulator, now to rebase
