+++ Detectors/TRD/simulation/CMakeLists.txt
@@ -14,13 +14,15 @@ o2_add_library(TRDSimulation
                        src/TRsim.cxx
                        src/Digitizer.cxx
                        src/TRDSimParams.cxx
+                       src/TrapSimulator.cxx
                PUBLIC_LINK_LIBRARIES O2::DetectorsBase O2::SimulationDataFormat O2::TRDBase)
 
 o2_target_root_dictionary(TRDSimulation
                           HEADERS include/TRDSimulation/Detector.h
                                   include/TRDSimulation/TRsim.h
                                   include/TRDSimulation/Digitizer.h
-                                  include/TRDSimulation/TRDSimParams.h)
+                                  include/TRDSimulation/TRDSimParams.h
+                                  include/TRDSimulation/TrapSimulator.h)
 
 if (OpenMP_CXX_FOUND)
     target_compile_definitions(${targetName} PRIVATE WITH_OPENMP)
